==> explore/Figure_workflow/AT_workflow/05_new_specifiy.Rmd <==
---
title: "05_new_specifiy"
output: html_document
---

```{r setup, include=FALSE}
library(tidyverse)
library(Seurat)
library(here)
```

```{r}
refquery <- readRDS(here('explore/RDS/07_df.RDS'))
fig3 <- subset(x = refquery, subset = cell_type_submain == "AT2" | cell_type_submain == "AT1")
fig3 <- saveRDS(fig3, file = "fig3.RDS")
```

```{r}
AT <- readRDS((here('explore/Figure_workflow/AT_workflow/fig3.RDS')))
```

```{r}
DimPlot(AT, group.by = "seurat_clusters", label = TRUE)
DimPlot(AT, group.by = "cell_type_submain", label = TRUE)
```

```{r}
AT@meta.data$new_submain <- ifelse(AT@meta.data$seurat_clusters %in% c(22), "AT1_1", "NA")
AT@meta.data$new_submain <- ifelse(AT@meta.data$seurat_clusters %in% c(10), "AT1_2", AT@meta.data$new_submain)

AT@meta.data$new_submain <- ifelse(AT@meta.data$seurat_clusters %in% c(29), "AT2_1", AT@meta.data$new_submain)
AT@meta.data$new_submain <- ifelse(AT@meta.data$seurat_clusters %in% c(33), "AT2_2", AT@meta.data$new_submain)
AT@meta.data$new_submain <- ifelse(AT@meta.data$seurat_clusters %in% c(3), "AT2_3", AT@meta.data$new_submain)
AT@meta.data$new_submain <- ifelse(AT@meta.data$seurat_clusters %in% c(17), "AT2_4", AT@meta.data$new_submain)
AT@meta.data$new_submain <- ifelse(AT@meta.data$seurat_clusters %in% c(14), "AT2_5", AT@meta.data$new_submain)
AT@meta.data$new_submain <- ifelse(AT@meta.data$seurat_clusters %in% c(21), "AT2_6", AT@meta.data$new_submain)
AT@meta.data$new_submain <- ifelse(AT@meta.data$seurat_clusters %in% c(28), "AT2_7", AT@meta.data$new_submain)
AT@meta.data$new_submain <- ifelse(AT@meta.data$seurat_clusters %in% c(2), "AT2_8", AT@meta.data$new_submain)


DimPlot(AT, group.by = "new_submain", label = TRUE)
```

```{r}
AT <- saveRDS(AT, file = "AT.RDS")
```


```{r}
DimPlot(AT, group.by = "Diagnosis")
```
